+++ Window.cpp
@@ -1,6 +1,6 @@
 #include "Window.h"
 
-Window::Window(int w, int h, std::string t) {
+Window::Window(GLint w, GLint h, std::string t) {
     width = w;
     height = h;
     title = t;
@@ -8,11 +8,11 @@ Window::Window(int w, int h, std::string t) {
 
 Window::Window() {}
 
-void Window::setWidth(int w) {
+void Window::setWidth(GLint w) {
     width = w;
 }
 
-void Window::setHeight(int h) {
+void Window::setHeight(GLint h) {
     height = h;
 }
 
@@ -20,7 +20,7 @@ void Window::setTitle(std::string t) {
     title = t;
 }
 
-void Window::setRGB(double r, double g, double b) {
+void Window::setRGB(GLfloat r, GLfloat g, GLfloat b) {
     red = r;
     green = g;
     blue = b;
